Merge r188931 - IconDatabase: syncThreadMainLoop() is unlocking m_syncLock twice when thread termination is requested https://bugs.webkit.org/show_bug.cgi?id=148429
Patch by Carlos Garcia Campos <cgarcia@igalia.com> on 2015-08-25
Reviewed by Filip Pizlo.

The lock is released an locked on every loop iteration, and then
unlocked again after the loop. There's an early break in the loop
when thread termination is requested that happens after the lock
is released but before is locked again, so that the unlock after
the loop is trying to unlock the lock again. This was not a
problem before, but the new Lock has an assertion to ensure that a
lock is not released twice.

* loader/icon/IconDatabase.cpp:
(WebCore::IconDatabase::syncThreadMainLoop): Clean up the thread
and return instead of breaking the loop when thread termination is
requested.

Canonical link: https://commits.webkit.org/166077.33@webkitgtk/2.9.92
git-svn-id: https://svn.webkit.org/repository/webkit/releases/WebKitGTK/webkit-2.10@188951 268f45cc-cd09-0410-ab3c-d52691b4dbfc